# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "$defs": {
          "ErrorDetail": {
            "properties": {
              "code": {
                "examples": [
                  "validation_error"
                ],
                "title": "Code",
                "type": "string"
              },
              "details": {
                "default": null,
                "title": "Details"
              },
              "message": {
                "examples": [
                  "Request validation failed"
                ],
                "title": "Message",
                "type": "string"
              },
              "request_id": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "default": null,
                "examples": [
                  "01J00000000000000000000000"
                ],
                "title": "Request Id"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "title": "ErrorDetail",
            "type": "object"
          }
        },
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "FrequencyType": {
        "enum": [
          "DAILY",
          "WEEKLY",
          "EVERY_X_HOURS",
          "AS_NEEDED"
        ],
        "title": "FrequencyType",
        "type": "string"
      },
...
      "ImagePurpose": {
        "enum": [
          "MEDICATION_LABEL",
          "PILL_REFERENCE",
          "VERIFICATION_IMAGE"
        ],
        "title": "ImagePurpose",
        "type": "string"
      },
...
      "ScheduleCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "default": true,
            "title": "Active",
            "type": "boolean"
          },
          "days_of_week": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "MON",
                    "TUE",
                    "WED",
                    "THU",
                    "FRI",
                    "SAT",
                    "SUN"
                  ],
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days Of Week"
          },
          "end_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "frequency_type": {
            "$ref": "#/components/schemas/FrequencyType"
          },
          "interval_hours": {
            "anyOf": [
              {
                "maximum": 168.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interval Hours"
          },
          "medication_id": {
            "format": "uuid",
            "title": "Medication Id",
            "type": "string"
          },
          "reminder_offset_minutes": {
            "default": 0,
            "maximum": 10080.0,
            "minimum": 0.0,
            "title": "Reminder Offset Minutes",
            "type": "integer"
          },
          "scheduled_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled Time"
          },
          "start_date": {
            "format": "date",
            "title": "Start Date",
            "type": "string"
          }
        },
        "required": [
          "medication_id",
          "frequency_type",
          "start_date"
        ],
        "title": "ScheduleCreateRequest",
        "type": "object"
      },
      "ScheduleResponse": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "format": "uuid",
            "title": "Created By",
            "type": "string"
          },
          "days_of_week": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days Of Week"
          },
          "end_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "frequency_type": {
            "$ref": "#/components/schemas/FrequencyType"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "interval_hours": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interval Hours"
          },
          "medication_id": {
            "format": "uuid",
            "title": "Medication Id",
            "type": "string"
          },
          "reminder_offset_minutes": {
            "title": "Reminder Offset Minutes",
            "type": "integer"
          },
          "scheduled_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled Time"
          },
          "start_date": {
            "format": "date",
            "title": "Start Date",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "medication_id",
          "user_id",
          "frequency_type",
          "scheduled_time",
          "days_of_week",
          "interval_hours",
          "start_date",
          "end_date",
          "reminder_offset_minutes",
          "active",
          "created_by",
          "created_at",
          "updated_at"
        ],
        "title": "ScheduleResponse",
        "type": "object"
      },
      "ScheduleUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "days_of_week": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "MON",
                    "TUE",
                    "WED",
                    "THU",
                    "FRI",
                    "SAT",
                    "SUN"
                  ],
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days Of Week"
          },
          "end_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "frequency_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FrequencyType"
              },
              {
                "type": "null"
              }
            ]
          },
          "interval_hours": {
            "anyOf": [
              {
                "maximum": 168.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interval Hours"
          },
          "reminder_offset_minutes": {
            "anyOf": [
              {
                "maximum": 10080.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reminder Offset Minutes"
          },
          "scheduled_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled Time"
          },
          "start_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          }
        },
        "title": "ScheduleUpdateRequest",
        "type": "object"
      },
      "SignedReadResponse": {
        "properties": {
          "expires_in_seconds": {
            "title": "Expires In Seconds",
            "type": "integer"
          },
          "image_id": {
            "format": "uuid",
            "title": "Image Id",
            "type": "string"
          },
          "read_url": {
            "title": "Read Url",
            "type": "string"
          }
        },
        "required": [
          "image_id",
          "read_url",
          "expires_in_seconds"
        ],
        "title": "SignedReadResponse",
        "type": "object"
      },
      "SignedUploadRequest": {
        "additionalProperties": false,
        "properties": {
          "content_type": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Content Type",
            "type": "string"
          },
          "patient_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Id"
          },
          "purpose": {
            "$ref": "#/components/schemas/ImagePurpose"
          },
          "size_bytes": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          }
        },
        "required": [
          "purpose",
          "content_type"
        ],
        "title": "SignedUploadRequest",
        "type": "object"
      },
      "SignedUploadResponse": {
        "properties": {
          "expires_in_seconds": {
            "title": "Expires In Seconds",
            "type": "integer"
          },
          "image_id": {
            "format": "uuid",
            "title": "Image Id",
            "type": "string"
          },
          "upload_fields": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Upload Fields",
            "type": "object"
          },
          "upload_url": {
            "title": "Upload Url",
            "type": "string"
          }
        },
        "required": [
          "image_id",
          "upload_url",
          "upload_fields",
          "expires_in_seconds"
        ],
        "title": "SignedUploadResponse",
        "type": "object"
      },
...
      "UserResponse": {
        "properties": {
          "accessibility_preferences": {
            "additionalProperties": true,
            "title": "Accessibility Preferences",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "full_name",
          "role",
          "timezone",
          "accessibility_preferences",
          "created_at",
          "updated_at"
        ],
        "title": "UserResponse",
        "type": "object"
      },
      "UserRole": {
        "enum": [
          "PATIENT",
          "CAREGIVER",
          "ADMIN"
        ],
        "title": "UserRole",
        "type": "string"
      },
      "UserUpdateRequest": {
        "properties": {
          "accessibility_preferences": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accessibility Preferences"
          },
          "full_name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "title": "UserUpdateRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Medication assistance API. It does not provide medical advice.",
    "summary": "Production API contract for MedGuide AI.",
    "title": "MedGuide AI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/schedules": {
      "get": {
        "operationId": "list_schedules_api_v1_schedules_get",
        "parameters": [
          {
            "in": "query",
            "name": "patient_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScheduleResponse"
                  },
                  "title": "Response List Schedules Api V1 Schedules Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Schedules",
        "tags": [
          "schedules"
        ]
      },
      "post": {
        "operationId": "create_schedule_api_v1_schedules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Schedule",
        "tags": [
          "schedules"
        ]
      }
    },
    "/api/v1/schedules/today": {
      "get": {
        "operationId": "list_todays_schedules_api_v1_schedules_today_get",
        "parameters": [
          {
            "in": "query",
            "name": "patient_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScheduleResponse"
                  },
                  "title": "Response List Todays Schedules Api V1 Schedules Today Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Todays Schedules",
        "tags": [
          "schedules"
        ]
      }
    },
    "/api/v1/schedules/{schedule_id}": {
      "delete": {
        "operationId": "delete_schedule_api_v1_schedules__schedule_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Schedule",
        "tags": [
          "schedules"
        ]
      },
      "patch": {
        "operationId": "update_schedule_api_v1_schedules__schedule_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Schedule",
        "tags": [
          "schedules"
        ]
      }
    },
    "/api/v1/uploads/signed-url": {
      "post": {
        "operationId": "create_signed_upload_url_api_v1_uploads_signed_url_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignedUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Signed Upload Url",
        "tags": [
          "uploads"
        ]
      }
    },
    "/api/v1/uploads/{image_id}": {
      "delete": {
        "operationId": "delete_uploaded_image_api_v1_uploads__image_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Image Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Uploaded Image",
        "tags": [
          "uploads"
        ]
      }
    },
    "/api/v1/uploads/{image_id}/signed-read-url": {
      "get": {
        "operationId": "create_signed_read_url_api_v1_uploads__image_id__signed_read_url_get",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Image Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedReadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Signed Read Url",
        "tags": [
          "uploads"
        ]
      }
    },
    "/api/v1/users/me": {
      "delete": {
        "operationId": "delete_me_api_v1_users_me_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Me",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "update_me_api_v1_users_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Me",
        "tags": [
          "users"
        ]
      }
    },
...
  },
...
}